requirements.txt
annotated-types==0.6.0
build==1.2.1
dill==0.3.8
email_validator==2.1.1
fastapi-cli==0.0.3
httptools==0.6.1
orjson==3.10.3
pydantic_core==2.18.2
python-dotenv==1.0.1
PyYAML==6.0.1
reflex-motion==0.0.1
reflex-simpleicons==0.0.5
Rx==3.2.0
tomlkit==0.12.4
twine==5.0.0
ujson==5.9.0
uvloop==0.19.0
wheel==0.43.0
